Extract `printStatementSequence`, `printMemberExpression`, and `printBlock` (#9723)
