+++ src/language-js/print/misc.js
@@ -3,11 +3,10 @@
 /** @type {import("assert")} */
 const assert = require("assert");
 const {
-  builders: { concat, softline, group, indent, join, line, hardline },
+  builders: { concat, group, indent, join, line, hardline },
 } = require("../../document");
 const {
   hasNewlineBetweenOrAfterDecorators,
-  isNumericLiteral,
   getParentExportDeclaration,
 } = require("../utils");
 
@@ -39,24 +38,6 @@ function printFunctionTypeParameters(path, options, print) {
     return path.call(print, "typeParameters");
   }
   return "";
-}
-
-function printMemberLookup(path, options, print) {
-  const property = path.call(print, "property");
-  const n = path.getValue();
-  const optional = printOptionalToken(path);
-
-  if (!n.computed) {
-    return concat([optional, ".", property]);
-  }
-
-  if (!n.property || isNumericLiteral(n.property)) {
-    return concat([optional, "[", property, "]"]);
-  }
-
-  return group(
-    concat([optional, "[", indent(concat([softline, property])), softline, "]"])
-  );
 }
 
 function printBindExpressionCallee(path, options, print) {
@@ -95,12 +76,24 @@ function printFlowDeclaration(path, printed) {
   return concat(["declare ", printed]);
 }
 
+function adjustClause(node, clause, forceSpace) {
+  if (node.type === "EmptyStatement") {
+    return ";";
+  }
+
+  if (node.type === "BlockStatement" || forceSpace) {
+    return concat([" ", clause]);
+  }
+
+  return indent(concat([line, clause]));
+}
+
 module.exports = {
   printOptionalToken,
   printFunctionTypeParameters,
-  printMemberLookup,
   printBindExpressionCallee,
   printTypeScriptModifiers,
   printDecorators,
   printFlowDeclaration,
+  adjustClause,
 };
